gemini.js: new file for using geminiAPI 	outSchema.js: new file for storing just the schema of output 	index.js: cleaned up
